src/cortex-api/requirements.txt
fastapi>=0.115.4,<0.116
uvicorn[standard]>=0.32.0,<0.34
pydantic>=2.9,<3
pydantic-settings>=2.6,<3
python-dotenv>=1.0,<2
sqlite-vec>=0.1.5,<0.2
mcp>=1.0,<2
sentence-transformers>=3.2,<4
ibm-watsonx-ai>=1.1,<2
llama-cpp-python>=0.3.2,<0.4
httpx>=0.27,<0.29
